requirements.txt
azure-common==1.1.28
azure-core==1.30.0
azure-identity==1.15.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-compute==30.5.0
azure-mgmt-core==1.4.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.0.1
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
idna==3.6
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
packaging==23.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pywin32==306
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.1
